+++ test/test_pytest_update_test_results.py
@@ -1,26 +1,39 @@
 from pathlib import Path
+from xml.etree import ElementTree as ET
 
 from _pytest.reports import TestReport
-from pytest_merge_xml.merge_xml import modify_xml
 from pytest_regressions.file_regression import FileRegressionFixture
+from pytest_update_test_results.update_test_results import modify_xml
 
 
-def test_xml_one_failed_two_passed(
-    datadir: Path, file_regression: FileRegressionFixture
-) -> None:
-    """Tests modify_xml() with list of previously failing flaky test
-    results that are now passing, and expects the xml to be overwritten
-    with the new result."""
-    #  'location', 'keywords', 'longrepr', and 'when'
+def test_one_failure(datadir: Path) -> None:
+    retest_results = {
+        "test_xml_three_failed_zero_passed": TestReport(
+            nodeid="test/test_pytest_merge_xml.py::test_xml_three_failed_zero_passed",
+            outcome="passed",
+            location=("test/test_pytest_merge_xml.py", 88, "test_xml_three_failed_zero_passed"),
+            keywords="",
+            longrepr=None,
+            when="call",
+        )
+    }
+
+    modified_file = datadir / "one_failure.retest.xml"
+    modify_xml(Path(datadir / "one_failure.xml"), retest_results, modified_file)
+
+    et = ET.parse(modified_file)
+    root_el = et.getroot()
+    testsuite_el = root_el.find("testsuite")
+    assert testsuite_el.attrib["failures"] == "0"
+    assert len(list(testsuite_el.iter("failure"))) == 0
+
+
+def test_two_failures(datadir: Path) -> None:
     retest_results_passed = {
         "test_xml_one_failed_two_passed": TestReport(
             nodeid="test/test_pytest_merge_xml.py::test_xml_one_failed_two_passed",
             outcome="passed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                9,
-                "test_xml_one_failed_two_passed",
-            ),
+            location=("test/test_pytest_merge_xml.py", 9, "test_xml_one_failed_two_passed"),
             keywords={},
             longrepr=None,
             when="call",
@@ -28,11 +41,7 @@ def test_xml_one_failed_two_passed(
         "test_xml_two_failed_one_passed": TestReport(
             nodeid="test/test_pytest_merge_xml.py::test_xml_two_failed_one_passed",
             outcome="passed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                49,
-                "test_xml_two_failed_one_passed",
-            ),
+            location=("test/test_pytest_merge_xml.py", 49, "test_xml_two_failed_one_passed"),
             keywords="",
             longrepr=None,
             when="call",
@@ -40,87 +49,71 @@ def test_xml_one_failed_two_passed(
         "test_xml_three_failed_zero_passed": TestReport(
             nodeid="test/test_pytest_merge_xml.py::test_xml_three_failed_zero_passed",
             outcome="failed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                88,
-                "test_xml_three_failed_zero_passed",
-            ),
+            location=("test/test_pytest_merge_xml.py", 88, "test_xml_three_failed_zero_passed"),
             keywords="",
             longrepr=None,
             when="call",
         ),
     }
-    obtained_file = datadir / "failure_tmp.xml"
-    modify_xml(Path(datadir / "failure.xml"), retest_results_passed, obtained_file)
-    file_regression.check(obtained_file.read_text(), extension=".xml")
+    modified_file = datadir / "failure_tmp.xml"
+    modify_xml(Path(datadir / "two_failures.xml"), retest_results_passed, modified_file)
+
+    root_el = ET.parse(modified_file).getroot()
+    testsuite_el = root_el.find("testsuite")
+    assert testsuite_el.attrib["failures"] == "1"
+
+    testcases = testsuite_el.findall("testcase")
+    assert testcases[1].find("failure") is None
+    assert testcases[2].find("failure") is not None
 
 
-def test_xml_two_failed_one_passed(
-    datadir: Path, file_regression: FileRegressionFixture
-) -> None:
-    """Tests modify_xml() with list of flaky test results, with some
-    still not passing, and expects the xml to be overwritten for the
-    passing tests."""
-    retest_results_some_passing = {
+def test_one_failure_one_error(datadir: Path) -> None:
+    """Tests modify_xml() with that change a XML with one failure and one error"""
+    retest_results_passed = {
         "test_xml_one_failed_two_passed": TestReport(
             nodeid="test/test_pytest_merge_xml.py::test_xml_one_failed_two_passed",
-            outcome="failed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                9,
-                "test_xml_one_failed_two_passed",
-            ),
-            keywords={},
-            longrepr=None,
-            when="call",
-        ),
-        "test_xml_two_failed_one_passed": TestReport(
-            nodeid="test/test_pytest_merge_xml.py::test_xml_two_failed_one_passed",
             outcome="passed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                49,
-                "test_xml_two_failed_one_passed",
-            ),
-            keywords="",
+            location=("test/test_pytest_merge_xml.py", 9, "test_xml_one_failed_two_passed"),
+            keywords={},
             longrepr=None,
             when="call",
         ),
         "test_xml_three_failed_zero_passed": TestReport(
             nodeid="test/test_pytest_merge_xml.py::test_xml_three_failed_zero_passed",
-            outcome="failed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                88,
-                "test_xml_three_failed_zero_passed",
-            ),
+            outcome="passed",
+            location=("test/test_pytest_merge_xml.py", 88, "test_xml_three_failed_zero_passed"),
             keywords="",
             longrepr=None,
             when="call",
         ),
     }
-    obtained_file = datadir / "failure_tmp.xml"
+    modified_file = datadir / "failure.retest.xml"
     modify_xml(
-        Path(datadir / "failure.xml"), retest_results_some_passing, obtained_file
+        Path(datadir / "one_failure_one_error.xml"),
+        retest_results_passed,
+        modified_file,
     )
-    file_regression.check(obtained_file.read_text(), extension=".xml")
+
+    root_el = ET.parse(modified_file).getroot()
+    testsuite_el = root_el.find("testsuite")
+    assert testsuite_el.attrib["failures"] == "0"
+    assert testsuite_el.attrib["errors"] == "0"
+    assert len(list(testsuite_el.iter("failure"))) == 0
+    assert len(list(testsuite_el.iter("error"))) == 0
 
 
-def test_xml_three_failed_zero_passed(
-    datadir: Path, file_regression: FileRegressionFixture
-) -> None:
-    """Tests modify_xml() with list of flaky test results, with some
-    still not passing, and expects the xml to be overwritten for the
-    passing tests."""
+def test_succeed_to_failure_on_retest_not_supported(datadir: Path) -> None:
+    """
+    For now, pytest-update-test-results is supposed to always run with `--last-failed` flag.
+
+    It convert failed outcomes into passed, but DOES NOT convert passed test cases into
+    failed ones (that could happen if `--update-xml` runs without `--last-failed`).
+    """
     retest_results_mix = {
         "test_xml_one_failed_two_passed": TestReport(
             nodeid="test/test_pytest_merge_xml.py::test_xml_one_failed_two_passed",
             outcome="failed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                9,
-                "test_xml_one_failed_two_passed",
-            ),
+            location=("test/test_pytest_merge_xml.py", 9, "test_xml_one_failed_two_passed"),
             keywords={},
             longrepr=None,
             when="call",
@@ -128,11 +121,7 @@ def test_xml_three_failed_zero_passed(
         "test_xml_two_failed_one_passed": TestReport(
             nodeid="test/test_pytest_merge_xml.py::test_xml_two_failed_one_passed",
             outcome="failed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                49,
-                "test_xml_two_failed_one_passed",
-            ),
+            location=("test/test_pytest_merge_xml.py", 49, "test_xml_two_failed_one_passed"),
             keywords="",
             longrepr=None,
             when="call",
@@ -140,16 +129,16 @@ def test_xml_three_failed_zero_passed(
         "test_xml_three_failed_zero_passed": TestReport(
             nodeid="test/test_pytest_merge_xml.py::test_xml_three_failed_zero_passed",
             outcome="failed",
-            location=(
-                "test/test_pytest_merge_xml.py",
-                88,
-                "test_xml_three_failed_zero_passed",
-            ),
+            location=("test/test_pytest_merge_xml.py", 88, "test_xml_three_failed_zero_passed"),
             keywords="",
             longrepr=None,
             when="call",
         ),
     }
-    obtained_file = datadir / "failure_tmp.xml"
-    modify_xml(Path(datadir / "failure.xml"), retest_results_mix, obtained_file)
-    file_regression.check(obtained_file.read_text(), extension=".xml")
+    modified_file = datadir / "all_failed.retest.xml"
+    modify_xml(Path(datadir / "no_failures.xml"), retest_results_mix, modified_file)
+
+    root_el = ET.parse(modified_file).getroot()
+    testsuite_el = root_el.find("testsuite")
+
+    assert testsuite_el.attrib["failures"] == "0"