+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
 
+0.2.0 (2022-10-03)
+------------------
+
+* Fix "error" test case outcomes not being updated (if they pass).
+* Do not rewrite the XML file if no changes on test outcomes were found.
+* Optimize the way XML results were being traversed.
+
 0.1.0 (2022-10-03)
 ------------------
 
-* First release on PyPI.
+* First release.